Add support for MMIO
The MemoryMapper class is added that listens on the data bus and
dispatches to one of many slaves that are mapped into the memory space.
Slaves can be either memory regions or MMIO devices.

The MmioDevice class can be subclassed to add new MMIO devices.
Registers can be mapped into the local memory space of devices in a
simple way by subclasses. The MmioDevice class handles mapping addresses
to actual registers.

Two simple devices are currently implemented:
- MachineTimers implements the mtime and mtimecmp registers. Note that
  the timer interrupt is not implemented yet.
- CharDev prints written characters to the console in a simulation.