Make sure rvfi.pc_wdata reliably reflects the next PC
We used to use the NEXT_PC pipeline register for pc_wdata. However, we
experienced a bug where NEXT_PC was correctly set for all instructions
but when the pipeline stalled, the instruction in IF would disappear due
to the PC still being updated. This meant that NEXT_PC didn't really
represent the next PC to be executed.

To fix this, we simply buffer one instruction's RVFI data and only send
it out when the next instruction is retired. This way, the next PC is
always accurate.