Execute MRET in WB
Since all CSR updates are performed in WB, executing MRET in EX as
before might lead to reading incorrect MEPC values. This is fixed for
now by executing MRET in WB but a better solutions should be found
(either some kind of hazard resolution as with normal registers or
somehow enforcing that all CSR accesses happen in the same stage).