Fix handling of same-cycle responses in MemBusControl
When writes where acknowledged in the same cycle as they where issued,
MemBusControl didn't notice this. This caused writes to be issued two
cycles in a row. This wasn't a problem for memories but devices executed
the same action twice because of this (e.g., writing a character twice
to the screen).