Re-evaluate Verilator model when the memory is updated
Whenever the memory updates inputs to the core, the core's Verilator
model should be re-evaluated. If not, its input signals appear to change
one time step later then expected. Although this was not a real issue
because multiple time steps are taken per cycle, it was confusing when
reading the wave form.