Add MemoryService to replace {I,D}BusService
This service is responsible for creating the external memory buses and
connecting internal buses to it. It also offers a way to "filter" the
data bus. This is done by allowing plugins to specify a callback to be
called whenever an internal data bus is created.

The service is currently implemented by the MemoryBackbone plugin. This
implementation allows for a single ibus and dbus as well as a single
dbus filter.