Allow Plugins to be constructed after Pipeline
Pipeline construction is now split in three steps: 1) construct
Pipeline, 2) add Plugins, and 3) call Pipeline::build().

This is a first step towards moving the Stage configuration out of
Pipeline. We don't want Plugins to statically add their logic to
specific Stages. Instead, the Stage they should use should be specified
when creating the Plugin. To make this easier, we want to separate the
construction of Pipeline from its configuration.