Create StaticPipeline class
This class contains all the code specific to a static pipeline: mainly
the creation and connection of the stages. The base Pipeline class only
contains code for managing plugins and services.

Plugins now also have to specify for what kind of Pipeline they work.
This is done through a generic type parameter of the base Plugin class.
All Plugins that only work for a StaticPipeline have been moved to the
riscv.plugins.scheduling.static package.