Make DataHazardResolver independent of the execute stage
It used the execute stage explicitly as the first stage that ever reads
a register value. The dependency was broken by passing this stage as a
constructor argument. However, this stage could be automatically
determined by registering which plugins write to arbitration.rsXNeeded
so this should be cleaned up at some point.