Add possibility to instantiate Pipeline without creating a Component
Pipeline is now a trait that does not inherit Component anymore.
Instead, it defines an abstract method called pipelineComponent that
should be implemented to return the Component in which the Pipeline
should be built.

This is done to allow the DynamicPipeline to use a StaticPipeline for a
subset of its stages without having those stages in a separate
Component.