Don't construct Pipeline.data in Pipeline's constructor
Since the construction of data needs Pipeline.config which may only be
defined in a subclass of Pipeline, this could lead to null-pointer
errors. Therefore, the class implementing Pipeline is now responsible
for providing the data object. This also allows a single data object to
be shared by multiple Pipelines (e.g., this will be used by
DynamicPipeline to share the data object with its statis issue
pipeline).