Store full-sized (CLEN bits) capabilities in the register file
Before, we packed version of capabilities (with only the bits that are
currently used) where stored. However, since the capability register
file can also be used to store CLEN-bit integers, this was incorrect.

The representation of capabilities is now implemented by multiple
classes to facilitate different representations needed at different
points in the microarchitecture:
- Capability: a trait for the capability fields;
- PackedCapability: all fields packed together in the smallest possible
  capability implementation;
- RegCapability: full, CLEN-bits, capability as stored in the register
  file.