Implement tagged memory and (basics of) LC and SC instructions
Tagged memory is currently implemented by storing a fixed amount of tag
bits on chip. The memory tagger is accessed through a new "capability
bus" which means that it should be easy to replace it with a more
realistic implementation later.

The basic functionality of the LC and SC instructions is implemented.
Access checks relative to DDC are not implemented yet.