Implement instruction prefetching
This is to allow IPC=1 for straight-line code even when there is an
access delay on the IBUS.

It is implemented by sending requests on the IBUS immediately when the
previous one is accepted (so we're not waiting for the response
anymore). When responses come in, we match them with the correct address
and send them to the fetch stage. They can optionally be buffered when
the pipeline is stalled.