Add retirement stage to the ooo pipeline
This stage can be used by plugins to add logic that should be run when
instructions are retired. This can move logic from the ROB (which
currently does everything related to retirement) to plugins.

As a first step, the existing RegisterFile plugin is added to the ooo
pipeline. This allows us to remove the special-purpose register file
that was currently used.