Refactor getService[T]
We now have service[T], which does the same as getService[T] previously,
and serviceOption[T] which returns an Option. This allows us to write
code that optionally depends on a service more ergonomically. For
example, we don't have to use a combination of hasService[T] and
getService[T] anymore, this can be done by matching on serviceOption[T].