Removed command to run performance experiments along build
I imagine that some people will not read carefully and just run the commands one after the other. In that case, we probably don't want them to run this one yet, so I removed it.

If a user really wants to run the performance experiments already, they can easily come up with the command themselves.